requirements.txt
amocrm-api==2.6.1
attrs==23.1.0
autoit==0.2.6
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
colorama==0.4.6
cssutils==2.7.1
currency-symbols==2.0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fonttools==4.41.1
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
lml==0.1.0
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoIt==0.6.5
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
PyJWT==2.8.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
texttable==1.6.7
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webcolors==1.13
webdriver-manager==3.8.6
wsproto==1.2.0
xlrd==2.0.1
xls2xlsx==0.2.0
xlwt==1.3.0
